Ignore `node_modules`
`node_modules` are development dependencies that differ per platform.
In particular, the committed files are for a Windows machine.
They don't work for a Linux machine.

Having them committed means developing on a different platform requires:
1. Deleting the committed files.
2. Installing the dependencies.
3. Remembering not to commit anything in the `node_modules` directory.

We can make it a smidge easier by untracking these files.
The workflow on all platforms becomes:
1. Install the dependencies.