Setup a `Makefile`
I don't know if this is cross-platform.
Hopefully, whatever doesn't work on Windows can be a small fix.
The idea isn't to write-off Windows nor make it second class,
more to track the dependency graph of the project explicitly.

We make an important change to the browserify command.
If we list the top-level entry point,
browserify will "do the right thing."
By that I mean, it will find all the dependencies.
It does this by following `require`s.

This change becomes helpful when we start to lean on npm more
for dependency management.