Try a top-down approach
We make PS effectively the entry-point to the program.
It takes the `htmlFunction`s and invokes them.
The rest of the program should work the same.

There is a bit of glue code with this approach.
Due to how FFI files work and require paths,
we take the PS output and the JS sources,
and massage the interfaces together.

As we need more JS modules in PS,
this glue module should be the place to add them.
When we take-over more functionality in PS,
this glue module shouldn't have to change.