Take over some of `htmlFunctions`
We arbitrarily choose to take the last three calls and move them to PS.
We could've started from the start and worked down.
Either way would ensure the same ordering.

There's nothing really special about what's happening here.
We're calling more JS code and letting it work how it works.

The PS is written to resemble the JS to help know show similarities.
It's not a requirement to stay this way.
In fact, it should most likely change once it's all converted.