Try a bottom-up approach
In this scenario, JS is calling out to PS.
It seems much simpler to go this way, and in some sense, it's easier.
But, we also carefully chose a very easily translatable function.

It might get harder when we start dealing with global mutable state.
But so long as we're doing things that are pure,
or that only affect the DOM, it will probably work out okay.

In any case, we try to make the PS look similar to the JS
so it's a bit easier to see what's similar.
We use `Fn1` instead of `(->)` since that's more descriptive.
`Fn1` is a synonym for `(->)`,
but it starts a pattern of using uncurried functions for JS interop.