Take over `calcFontPoint`
This function differs in behavior from JS implementation.
The reason it's different is that we're forced to handle array bounds.
If we get an array with zero or one element,
we return the original array.

In the JS implementation,
it silently converts the missing elements to `NaN`.
If the JS implementation is dependent on converting to `NaN`s,
and cannot work with the original array,
this function should change to match that.
It doesn't seem likely, but it's possible.

What would be better is if we weren't using arrays as makeshift tuples.
If we need two numbers, pass the two numbers.
Of course, all of this is predicated on the values being not-null.

We also add in `pcbfont`.
`drawText` was expected to exist, and would fail if it wasn't around.