Take over `drawtext`
There's nothing great about what we wrote.

It's almost a direct copy from JS.
There were a couple of cases where we had to change the logic a bit.
They were all cases of exiting gracefully if something goes wrong,
instead of throwing an exception.

Now that it's in PS, it can be made more idiomatic.