Build the line one go
We were making the line by draw each individual part of the line.
Since the line is all connected, this should be equivalent.

Moving to this style where we don't have to think about array indices
means we can also not think about array bounds.
We can speak more declaratively about what we want to do.

It's important to note that while it might look like we've
increased the complexity (we went from three `for_`s to five `for_`s),
`for_` doesn't work the same way as it does in JS.
`for_` is polymorphic in the value it operates on.

The outer `for_` is operating on `Array CodePoint`.
That iterates the array like in JS.
The second `for_` is operating on `Maybe (Array (Array Point))`.
That cases on the two possible values of `Maybe _`
(`Nothing` and `Just _`) and calls the function if it's `Just _`.
This run in constant time.
The third `for_` is like the first, but on `Array (Array Point)`.
The fourth `for_` is like the second,
but on `Maybe { head :: Point, tail :: Array Point }`.
The fifth `for_` is like the first and third, but on `Array Point`.

We've not increased the complexity at all.
In fact,
we probably decreased it as we're not going through `MaybeT _ _`.